+++ render.h
@@ -75,14 +75,14 @@ typedef struct
 	vrect_t		vrect;				// subwindow in video for refresh
 									// FIXME: not need vrect next field here?
 	vrect_t		aliasvrect;			// scaled Alias version
-	int			vrectright, vrectbottom;	// right & bottom screen coords
-	int			aliasvrectright, aliasvrectbottom;	// scaled Alias versions
+	long		vrectright, vrectbottom;	// right & bottom screen coords
+	long		aliasvrectright, aliasvrectbottom;	// scaled Alias versions
 	float		vrectrightedge;			// rightmost right edge we care about,
 										//  for use in edge list
 	float		fvrectx, fvrecty;		// for floating-point compares
 	float		fvrectx_adj, fvrecty_adj; // left and top edges, for clamping
-	int			vrect_x_adj_shift20;	// (vrect.x + 0.5 - epsilon) << 20
-	int			vrectright_adj_shift20;	// (vrectright + 0.5 - epsilon) << 20
+	long		vrect_x_adj_shift20;	// (vrect.x + 0.5 - epsilon) << 20
+	long		vrectright_adj_shift20;	// (vrectright + 0.5 - epsilon) << 20
 	float		fvrectright_adj, fvrectbottom_adj;
 										// right and bottom edges, for clamping
 	float		fvrectright;			// rightmost edge, for Alias clamping
